test(persistence): stabilize public surface fixtures
